requirements.txt
altair==5.0.1
anyio==3.7.1
asyncstdlib==3.10.8
attrs==23.1.0
base58==2.1.1
black==23.7.0
blinker==1.6.2
cachetools==4.2.4
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.6
colorama==0.4.6
construct==2.10.68
construct-typing==0.5.6
decorator==5.1.1
gitdb==4.0.10
GitPython==3.1.32
h11==0.14.0
httpcore==0.16.3
httpx==0.24.1
idna==3.4
importlib-metadata==6.8.0
iniconfig==2.0.0
Jinja2==3.1.2
jsonalias==0.1.1
jsonschema==4.18.4
jsonschema-specifications==2023.7.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
mypy-extensions==1.0.0
numpy==1.25.1
packaging==23.1
pandas==2.0.3
pathspec==0.11.1
Pillow==9.5.0
platformdirs==3.9.1
pluggy==1.2.0
protobuf==4.23.4
pyarrow==12.0.1
pycparser==2.21
pydeck==0.8.0
Pygments==2.15.1
Pympler==1.0.1
PyNaCl==1.5.0
pytest==7.4.0
python-dateutil==2.8.2
python-telegram-bot==20.4
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
referencing==0.30.0
requests==2.31.0
rfc3986==1.5.0
rich==13.4.2
rpds-py==0.9.2
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
solana==0.30.2
solders==0.18.1
streamlit==1.25.0
tenacity==8.2.2
toml==0.10.2
toolz==0.12.0
tornado==6.3.2
types-cachetools==4.2.10
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==4.3.1
urllib3==2.0.4
validators==0.20.0
websockets==11.0.3
zipp==3.16.2
